Mount Fuji
Mount Fuji, Japanese Fuji-san, also spelled Fujisan, also called
Fujiyama or Fuji no Yama,  highest mountain in Japan. It rises to
12,388 feet (3,776 metres) near the Pacific Ocean coast in Yamanashi
and Shizuoka ken (prefectures) of central Honshu, about 60 miles (100
km) west of the Tokyo-Yokohama metropolitan area. It is a volcano that
has been dormant since its last eruption, in 1707, but is still
generally classified as active by geologists. The mountain is the major
feature of Fuji-Hakone-Izu National Park (1936), and it is at the
centre of a UNESCO World Heritage site designated in 2013.